+++ src/main/kotlin/net/fallingangel/jimmerdto/psi/element/DTOAnnotationParameter.kt
@@ -1,12 +1,12 @@
 package net.fallingangel.jimmerdto.psi.element
 
 import com.intellij.psi.PsiElement
-import net.fallingangel.jimmerdto.psi.mixin.DTOElement
+import net.fallingangel.jimmerdto.psi.mixin.DTONamedElement
 
-interface DTOAnnotationParameter : DTOElement {
+interface DTOAnnotationParameter : DTONamedElement {
     val name: PsiElement
 
     val eq: PsiElement
 
-    val value: DTOAnnotationValue
+    val value: DTOAnnotationValue?
 }