+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+* Issue [#55](https://github.com/ClearPlume/jimmer-dto/issues/55): Add semantic check and quick fixes to DTO modifiers
+
 ### Deprecated
 
 * Deprecate 'CreateOrJumpToJimmerDtoFile'; to be removed in 0.0.8